Add load_placement and load_config methods...
Add `load_placement` and `load_config` methods to `PlacementController`
class.  Catch exceptions in `apply_swap_results()` method and call
`load_placement()` to refresh the placement by requesting the entire
grid contents from the remote `ModifierController`.  Note that passing
`true` as the first parameter to `load_placement`, `load_config` will be
called after syncing the placement.  If the argument is not supplied or
is set to`false`, `load_config` will not be called.