Add previous/next methods to PlacementController
Add `previous` and `next` methods to `PlacementController` class.  These
methods step backwards and forwards through the saved `SwapContext`
states.  If the last `SwapContext` in the
`PlacementController.swap_contexts` list is active when `next` is
called, create a new `SwapContext` and request an iteration from the
remote `ModifierController`.

Add `goto`, `home` and `end` methods to `PlacementController` class.
The `goto` method takes the index of the desired `SwapContext` instance
in the `PlacementController.swap_contexts` array.  The `home` and `end`
methods change focus to the first and last `SwapContext` instances,
respectively.

Assign the following shortcut keys:
    `home` -> `PlacementController.home`
    `end` -> `PlacementController.end`
    `space` -> `PlacementController.next`
    `n` -> `PlacementController.next`
    `right` -> `PlacementController.next`
    `p` -> `PlacementController.previous`
    `left` -> `PlacementController.previous`