Add Net class, draw net links to connected blocks
Add utility members to `Net` class to visually highlight the net using a circle
to denote the center-of-gravity of the net, and lines connecting each block of
the net to the center-of-gravity.  Currently these net representations are not
drawn automatically (i.e., they're only drawn by manually calling the
`highlight` method).

Add list of net ids to "Current block" information table.  This should help debugging discrepancies between the list of
net ids in conflicting swap from<->to/to<->from results.