Fix block position updating
Block positions were being set immediately based on a `SwapContext` that was
just created.  This meant that the `SwapContext` did not contain any swaps yet,
and, thus, no change was being made to the block positions.  As of this commit,
the `SwapContext` is not applied until after a `swaps_end` status update has
been received.