Add manager_comparator.html, compare 2 placements
Add `manager_comparator.html`, which displays 2 placement grids.  Currently,
the grids must be updated manually from the console, or using a hard-coded
address of 'tcp://maeby.fobel.net:12346'.  However, the idea is that multiple
placement managers can be connected to, and a placement can be loaded from any
of these placement managers.  Furthermore, swap contexts may be loaded (if
available), which may be used to render block and swap-link formatting.

N.B. Swap context block and swap-link formatting currently is applied using
`attr('fill', ...)`, etc.  This causes any base CSS formatting to be
overridden.  This should be changed to instead apply (a) new CSS class(es) to
the corresponding blocks and/or swap-links.