Use `gzip` arg in PlacementManager.get_result
Use `gzip` command field in `PlacementManager.get_result`, which requests the
remote `PlacementManager` to send the `result` field of the response encoded in
`gzipped` `base64`.

The main reason for this is to help speed up the `get_swap_context()` method.
Since even for a smallish netlist (`ex5p`), a single JSON-encoded list of all
swap infos can reach over 1MB.  Since the placement manager could easily be on
a remote computer, gzipping the swap context first should speed things up by
about 10x.

N.B., the call to `translate_value` in the Python `PlacementManager` is still a
little slow, so there will still be a delay of up to about a second.