[WIP] Add skeleton class for PlacementManagerGrid
Add skeleton class (with docs) for `PlacementManagerGrid`.  The main idea here
is to add controls to select a placement, swap context, and place
configuration, based on a corresponding key value.

See #9.

Currently, the `refresh_keys` method is implemented, which requests the
placement and swap context keys available from the remote placement manager,
and saves the union of the two sets of keys.  This list of keys is then saved
as the current list of available placement keys from the remote placement
manager.

N.B. Currently, requesting a swap context for a key, `k_i`, where a swap
context does not exist, but a placement does exist for the same key, results in
an error.  However, the placement manager could be modified to automatically
generate a random swap context consisting of a set of swaps that would
transform placement `k_{i}` to placement `k_{i + 1}` (there should be many,
many such sequences of swaps).  This could be used to simply "fill in" missing
swap contexts, or could possibly be used at the CoffeeScript level as a means
of comparing/generating intermediate placements from two different placement
managers, possibly tied to different algorithms. See #10.

Refactor `PlacementGrid` template creation to be performed in the new
`get_templates` method.  This method returns a dictionary-mapping between each
template label and the corresponding Underscore.js template object.  This
allows sub-classes to easily add new templates, as well as customize the
default header template.

Add `PlacementManagerComparator` class, which requires a `PlacementManager`
instance to be passed to the `reset_grid_[ab]` methods, which makes the
creation of a `PlacementManagerGrid` possible.