[NB] Add `save` method to grid class
The `save` method uses a data URI to direct the browser to save the SVG
rendering of the current grid state, including block and swap link formats.
Note that it is unfortunately not (currently) possible to set the filename of
the dynamically created file.  Also, there is currently no UI to
trigger/request a save, but saving can easily be done in the browser's
Javascript console (at least with Chrome).