Add area range id labels on grid, see #5
Add area range id labels on grid.  Currently, the ids are always visible.  The
style of the labels can be adjusted using the 'area_range-label' CSS class.
Note that the size of the font is set programmatically, based on the grid's
`scale`, such that the font is scaled based on the grid block sizes.